+++ source/databricks/calculation_engine/package/calculation/energy/resolution_transition_factory.py
