+++ source/databricks/calculation_engine/package/calculation/output/energy_storage_model_factory.py
@@ -18,7 +18,7 @@ from pyspark.sql.types import DecimalType
 
 from package.calculation.calculator_args import CalculatorArgs
 from package.calculation.energy.data_structures.energy_results import EnergyResults
-from package.calculation.energy.quarterly_resolution_transition_factory import (
+from package.calculation.energy.resolution_transition_factory import (
     get_resolution,
 )
 from package.calculation.output.add_meta_data import add_metadata
